python-neo4j-driver/requirements.txt
neo4j==1.7.5
neobolt==1.7.15
neotime==1.7.4
pytz==2019.3
PyYAML==5.1
six==1.11.0
pytest==5.0.1
